backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.6.0
requests>=2.31.0
